✨ (new): Add an alert window to display when user tapped on any flag buttons
Created a new method "askQuestion()" to be called when user tap on the continue button in the alert window. This resets the game state by randomizing the order of the countries and the correctAnswer property number. The alert window is set to be display everytime a user tapped on any of the flag buttons on screen.