✨ (new): Add tracking of user scores and new alert displays
Added new @State properties for tracking user scores and diplaying of alert messages based on the user's in-game answer selection. Implemented feature to capped each round to a total of maximum 8 questions before having the user to reset their scores. A display of the current number of questions have also been implemented for user to keep track.